Added payment app moved payment code to new app. Ran tests individually and passed.
