+++ wpa_project/payment/models/__init__.py
@@ -1,0 +1,2 @@
+from .payment_log import PaymentLog
+from .refund_log import RefundLog
