+++ wpa_project/payment/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .payment_serializers import PaymentSerializer
