+++ wpa_project/payment/tests/tests_process_payment.py
@@ -2,10 +2,11 @@ import logging
 import uuid
 from django.test import TestCase, Client
 from django.urls import reverse
+from django.conf import settings
+from django.apps import apps
 
-from ..models import ClassRegistration, PaymentLog, User
-from ..src import SquareHelper
-
+# from ..models import  User
+from ..models import PaymentLog
 logger = logging.getLogger(__name__)
 
 
@@ -14,11 +15,12 @@ class TestsProcessPayment(TestCase):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        self.User = apps.get_model(app_label='student_app', model_name='User')
 
     def setUp(self):
         # Every test needs a client.
         self.client = Client()
-        self.test_user = User.objects.get(pk=2)
+        self.test_user = self.User.objects.get(pk=2)
         self.client.force_login(self.test_user)
         session = self.client.session
         session['idempotency_key'] = str(uuid.uuid4())
@@ -28,12 +30,12 @@ class TestsProcessPayment(TestCase):
 
     def test_get_payment_page(self):
         # Get the payment page
-        response = self.client.get(reverse('registration:process_payment'), secure=True)
+        response = self.client.get(reverse('payment:process_payment'), secure=True)
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed('student_app/square_pay.html')
 
     def test_bypass(self):
-        response = self.client.post(reverse('registration:process_payment'),
+        response = self.client.post(reverse('payment:process_payment'),
                                     {'bypass': ''}, secure=True)
         self.assertTemplateUsed('student_app/message.html')
         pl = PaymentLog.objects.all()
@@ -45,7 +47,7 @@ class TestsProcessPayment(TestCase):
         session['line_items'] = [{'name': 'Class on None student id: 1',
                                   'quantity': '1', 'base_price_money': {'amount': 0, 'currency': 'USD'}}]
         session.save()
-        response = self.client.post(reverse('registration:process_payment'),
+        response = self.client.post(reverse('payment:process_payment'),
                                     {'bypass': ''}, secure=True)
         self.assertTemplateUsed('student_app/message.html')
         pl = PaymentLog.objects.all()
