+++ wpa_project/payment/views/__init__.py
@@ -1,0 +1,2 @@
+from .payment_view import PaymentView
+from .process_payment_view import ProcessPaymentView
