+++ wpa_project/student_app/serializers/__init__.py
@@ -1,4 +1,3 @@
-from .payment_serializers import PaymentSerializer
 from .student_family_serializers import StudentFamilySerializer
 from .student_serializers import StudentSerializer
 from .theme_serializers import ThemeSerializer
