+++ wpa_project/student_app/src/__init__.py
@@ -1,3 +1,1 @@
 from .class_registration_helper import ClassRegistrationHelper
-from .email import EmailMessage
-from .square_helper import SquareHelper
