+++ wpa_project/student_app/temp_tests.py
@@ -7,9 +7,9 @@
 # from .tests.tests_email import TestsEmail
 # from .tests.tests_index import TestsIndex
 # from .tests.tests_payment import TestsPayment
-# from .tests.tests_process_payment import TestsProcessPayment
+# from payment.tests.tests_process_payment import TestsProcessPayment
 # from .tests.tests_search import TestsClassSearch
-# from .tests.tests_square_helper import TestsSquareHelper
+# from payment.tests.tests_square_helper import TestsSquareHelper
 # from .tests.tests_student import TestsStudent
 # from .tests.tests_student import TestsStudentAPI
 # from .tests.tests_student_family import TestsStudentFamily
@@ -18,31 +18,31 @@
 # from .tests.tests_student_registration import TestsRegisterStudent
 # from .tests.tests_theme import TestsTheme
 # from .tests.tests_unregister import TestsUnregisterStudent
-# from .tests.tests_unregister import TestsUnregisterStudent2
+from .tests.tests_unregister import TestsUnregisterStudent2
 
-import json
-import logging
-import time
-import uuid
-from django.db.models import Q
-from django.core import mail
-from django.test import TestCase, Client
-from django.urls import reverse
-
-from .src import EmailMessage
-from .models import BeginnerClass, ClassRegistration, Student, StudentFamily, PaymentLog, User
+# import json
+# import logging
+# import time
+# import uuid
+# from django.db.models import Q
+# from django.core import mail
+# from django.test import TestCase, Client
+# from django.urls import reverse
 #
-logger = logging.getLogger(__name__)
-
-
-class TestsTemp(TestCase):
-    fixtures = ['f1']
-
-    def setUp(self):
-        # Every test needs a client.
-        self.client = Client()
-        self.test_user = User.objects.get(pk=1)
-        self.client.force_login(self.test_user)
+# from .src import EmailMessage
+# from .models import BeginnerClass, ClassRegistration, Student, StudentFamily, PaymentLog, User
+# #
+# logger = logging.getLogger(__name__)
+#
+#
+# class TestsTemp(TestCase):
+#     fixtures = ['f1']
+#
+#     def setUp(self):
+#         # Every test needs a client.
+#         self.client = Client()
+#         self.test_user = User.objects.get(pk=1)
+#         self.client.force_login(self.test_user)
         # self.email_dict = {'total': 5, 'receipt': 'https://example.com',
         #                    'line_items': [{'name': 'Class on None student id: 1',
         #                                    'quantity': '1', 'base_price_money': {'amount': 500, 'currency': 'USD'}}]}
