+++ wpa_project/student_app/templates/student_app/square_pay.html
@@ -92,9 +92,10 @@
     Payments are processed with Square. Your credit card information is not stored on our servers.
   </div>
   <div class="row">
-      <form class="col" id="payment-form" novalidate action="/process_payment" method="post">
+      <form class="col" id="payment-form" novalidate action="/payment/process_payment" method="post">
         {% csrf_token %}
-        <input type="hidden" id="sq_token" name="sq_token" url="{% url 'registration:payment' %}" value="">
+        <input type="hidden" id="sq_token" name="sq_token" value="">
+        <input type="hidden" id="pay_url" value="{% url 'registration:class_payment' %}">
         <div id="card-container"></div>
         <button class="btn btn-primary m-2" id="card-button" type="button"> Pay Now </button>
       </form>
