+++ wpa_project/student_app/tests/tests_payment.py
@@ -4,8 +4,9 @@ import uuid
 from django.test import TestCase, Client
 from django.urls import reverse
 
-from ..models import BeginnerClass, ClassRegistration, PaymentLog, Student, User
-from ..src import SquareHelper
+from ..models import BeginnerClass, ClassRegistration, Student, User
+# from payment.src import SquareHelper
+from payment.models import PaymentLog
 
 logger = logging.getLogger(__name__)
 
@@ -42,22 +43,22 @@ class TestsPayment(TestCase):
 
     def test_payment_success(self):
         # process a good payment
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-ok'}, secure=True)
         self.eval_content(json.loads(response.content), 'COMPLETED', [], False, 1)
 
     def test_payment_card_decline(self):
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-declined'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'Card was declined, ', True, 0)
 
     def test_payment_card_bad_cvv(self):
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-rejected-cvv'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'Error with CVV, ', True, 0)
 
     def test_payment_card_bad_expire_date(self):
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-rejected-expiration'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'Invalid expiration date, ', True, 0)
 
@@ -65,7 +66,7 @@ class TestsPayment(TestCase):
         session = self.client.session
         del session['idempotency_key']
         session.save()
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-ok'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'payment processing error', False, 0)
 
@@ -76,19 +77,19 @@ class TestsPayment(TestCase):
                          secure=True)
         ik = self.client.session['idempotency_key']
 
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-rejected-cvv'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'Error with CVV, ', True, 0)
         cr = ClassRegistration.objects.get(pk=1)
         self.assertNotEqual(ik, cr.idempotency_key)
 
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-rejected-cvv'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'Error with CVV, ', True, 0)
         cr = ClassRegistration.objects.get(pk=1)
         self.assertNotEqual(ik, cr.idempotency_key)
 
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-rejected-expiration'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'Invalid expiration date, ', False, 0)
         cr = ClassRegistration.objects.get(pk=1)
@@ -98,13 +99,13 @@ class TestsPayment(TestCase):
         session = self.client.session
         del session['line_items']
         session.save()
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'sq_token': 'cnon:card-nonce-ok'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'missing line items.', False, 0)
 
     def test_payment_invalid_post(self):
         session = self.client.session
         session.save()
-        response = self.client.post(reverse('registration:payment'),
+        response = self.client.post(reverse('registration:class_payment'),
                                     {'token': 'cnon:card-nonce-ok'}, secure=True)
         self.eval_content(json.loads(response.content), 'ERROR', 'payment processing error', False, 0)
