+++ wpa_project/student_app/urls.py
@@ -8,6 +8,7 @@ urlpatterns = [
         path('beginner_class/<int:beginner_class>/', BeginnerClassView.as_view(), name='beginner_class'),
         path('beginner_class', BeginnerClassView.as_view(), name='beginner_class'),
         path('class_list', BeginnerClassListView.as_view(), name='class_list'),
+        path('class_payment', PaymentView.as_view(), name='class_payment'),
         path('class_registered_table', ClassRegisteredTable.as_view(), name='class_registered_table'),
         path('class_registration/<int:reg_id>', ClassRegistrationView.as_view(), name='class_registration'),
         path('class_registration', ClassRegistrationView.as_view(), name='class_registration'),
@@ -15,10 +16,9 @@ urlpatterns = [
         path('class_status/<str:class_id>/', ClassStatusView.as_view(), name='class_status'),
         path('costs/<int:cost_id>', CostsView.as_view(), name='costs'),
         path('costs', CostsView.as_view(), name='costs'),
-        path('payment', PaymentView.as_view(), name='payment'),
         path('privacy', PrivacyView.as_view(), name='privacy'),
-        path('process_payment', ProcessPaymentView.as_view(), name='process_payment'),
-        path('process_payment/<str:message>/', ProcessPaymentView.as_view(), name='process_payment'),
+        # path('process_payment', ProcessPaymentView.as_view(), name='process_payment'),
+        # path('process_payment/<str:message>/', ProcessPaymentView.as_view(), name='process_payment'),
         path('profile', ProfileView.as_view(), name='profile'),
         path('search', SearchView.as_view(), name='search'),
         path('student_api/<int:student_id>/', StudentApiView.as_view(), name='student_api'),
