+++ wpa_project/student_app/views/costs_view.py
@@ -1,7 +1,7 @@
 import logging
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import render
-from django.http import  HttpResponseForbidden
+from django.http import HttpResponseForbidden
 from django.views.generic.base import View
 from django.shortcuts import get_object_or_404
