+++ wpa_project/student_app/views/student_list_json.py
@@ -13,7 +13,7 @@ from django.utils import timezone
 
 from ..forms import BeginnerClassForm, ClassAttendanceForm
 from ..models import Student, BeginnerClass, ClassRegistration, CostsModel
-from ..src import SquareHelper
+# from payment.src import SquareHelper
 
 logger = logging.getLogger(__name__)
