+++ wpa_project/student_app/views/unregister_view.py
@@ -9,7 +9,7 @@ from rest_framework.views import APIView
 from rest_framework import permissions
 from rest_framework.response import Response
 from ..serializers import UnregisterSerializer
-from ..src import EmailMessage, SquareHelper
+from payment.src import EmailMessage, SquareHelper
 from ..models import BeginnerClass, ClassRegistration, StudentFamily
 logger = logging.getLogger(__name__)
