+++ wpa_project/wpa_project/urls.py
@@ -22,6 +22,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/', include('allauth.urls')),
     path('', include('student_app.urls', namespace='registration')),
+    path('payment/', include('payment.urls', namespace='payment')),
 ]
 if settings.DEBUG:
     urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
